Moved files to root
Awesome work with the prototype Corey. Corey's prototype was pushed in a separate folder so I just moved the files to the root of repository. Everyone make sure when you open the project in Godot it's to the root of the repo, and not in a subfolder. I also changed to the project name so it'll show up in Godot as Never On Time.