Changed renderer to compatibility and tweaked player physics
My linux machine kept crashing trying to run the game with the forward+ renderer so I had to change it to compatibility. This shouldn't affect anything and just makes it so the game can run on more computers. Also tweaked some numbers in the player script to make it less floaty and make sure you can actually fall in 1 tile wide gaps.