Moved past level down, added time shift distance, added selective platform and put it in all ladder scenes, started work on sprint button and checkpoints
the selective platforms solve some ladder issues, and cause some others, but they feel pretty good so far. I thought that platforming might be a little hard at such a high speed and small sprite so i was gonna implement a quick sprint function. Also renamed all the groups to be consistent. I went through most of the scripts and made sure to rename things but if something is broken that worked before check for a is_in_group() with a lower case group name. Yell at me if there's issues or you don't like the sprint idea.